Add MPI explicitly for netcdf4
This is needed for the hdf5.h, which depends on mpi.h. A better solution would
be to make hdf5 automatically pull in the MPI dependency, but at the moment
even if you put:

run: [mpi]

into hdf5.yaml, the CPPFLAGS/CFLAGS will not be specified for the netcdf4
package, thus the build fails.