Refactor Parmetis-CMake builds/dynamic link flags
Parmetis is really managed by a CMake build, and we need access to the
underlying CMake commands to inject specific build linking flags, as
needed on supercomputers such as the Cray XE6.

I've also added the ability to pass DYNAMIC_EXE_LINKER_FLAGS in to
the CMake build, an example flag needed for dynamic linking on the
Cray.