Refactor Python packages
Most of our Python packages are importing from distutils_package, not
python_package. This is probably a good thing, as I don't really want to
deal with all the complexity that .eggs bring us with comparatively
little return.

[1] Explicitly add the setuptools dependency to anything importing from
setuptools (such as our python_package base recipe).  As I've mentioned
in the past, distribute is has been merged into setuptools and is now
deprecated, and the dependency on setuptools should be encoded in the
base package.

[2] Rename the base packages to:

- distutils_package (generic, uses native distutils installer)
- setuptools_package (inherits and overrides distutils_package, uses
setuptools installer)

Action: Rename python_package -> setuptools_package.

[3] Remove the distribute package completely from the repository.

[4] Set any Python package that doesn't rely on setuptools to install to
use distutils_package.

We should still do an audit of our current package base for this.