ENH: autotools/configure handling w/ env append
Add the ability to append to environment variables in the configure
stage.  This is especially important if the user needs to adjust CPFLAGS
or LDFLAGS, which would normally be overridden if specified in the
configure line.

This also adds two in-hashstack test packages.  These should eventually
be moved into a different directory, perhaps test/pkgs.