DIRTY/WIP - BG/Q Port
This is currently crashing on numpy tests.

The problem occurs in test_datetime but doesn't seem to be related
to the numpy/datetime library.  Datetime was modified in 1.7, and
its possible this code has bugs in it specific to the PowerPC
architecture.  The next thing to do is to try and back off to
numpy 1.6