gnuradio: initial add of 3.6.5
This commit includes gtk2 and a minimal set of dependencies
needed to build pygtk. gnuradio core is built, but the more useful
gnuradio-companion has not been built due to minor issues with pygtk
needing canberra-gtk-module (xdg sound and friends)

At least there packages have been added

    atk.yaml
    cairo.yaml
    fontconfig.yaml
    gdk-pixbuf.yaml
    glib2.yaml
    gnuradio.yaml
    gobject-introspection.yaml
    gtk2.yaml
    libcanberra.yaml
    libvorbis.yaml
    pango.yaml
    pixman.yaml
    poppler.yaml
    pycairo.yaml
    pygobject.yaml
    pygtk.yaml

gnuradio could be updated to a newer version, however a number of
dependencies may need to be upgraded before this may be possible.

Signed-off-by: Jimmy Tang <jcftang@gmail.com>