Git: depend on Perl
This is necessary so that the git build system checks our own Perl for the
presence of the Error.pm module, and since it doesn't exists, installs it.
Without the Perl dependency, it checks the system-wide Perl installation, which
sometimes containst he Error.pm module, so git doesn't install it and then it
is missing at runtime.