ca-bundle: install a generated ca-cert bundle for curl
On OSX git calls curl from hashstack, it will fail on https based
repositories if SSL verification is left to the default value of
true.

This commit pulls in a generated ca cert bundle generated from with
the mk-ca-bundle.pl that ships with curl.

It is probably better to ship a semi-regularly updated ca-bundle
than to tell users to turn off verification that verification of
ssl based services are broken.

Signed-off-by: Jimmy Tang <jcftang@gmail.com>