Add CPPFLAGS to CMake explicitly
Even though CMake is supplied with (via `base/cmake_package.py`)

-DCMAKE_PREFIX_PATH="${BOOST_DIR};${BZIP2_DIR};..."

For some reason, some of the `-I` statements do not get propagated to g++. If
we add the line

-DCMAKE_CXX_FLAGS:STRING="${CPPFLAGS}"

Then all the necessary `-I` statements are propagated. Note the quotation marks
around "${CPPFLAGS}$, those are necessary, since there are spaces in the
variable.