set OPENBLAS_DIR in openblas.yaml when build dependencies
I found that the numpy package is not able to find openblas at compile time, because it is looking in `$OPENBLAS_DIR/lib` and `OPENBLAS_DIR` is not set in `openblas.yaml `. I think that this setting should be unable. At least with this change numpy finds openblas for me.