CMake: use our zlib, bzip2 and curl
This way CMake does not build its own versions, but rather uses Hashstack
versions. We also made sure the proper RPATH will be used in the `cmake`
executable via the CMAKE_PREFIX_PATH option.

Only three other libraries are still build as part of CMake: expat, jsoncpp,
libarchive, so we commented out options that disable them for now.

As a bonus, this patch also fixes a problem on OSX 10.10 where the curl
component of CMake fails to build with gcc (we now use Hashstack's version that
builds).