CMake: set OSX specific variables properly
On OSX 10.10, one gets errors of the type:

[cmake] CMake Error at Modules/Platform/Darwin.cmake:76 (message):
[cmake]   CMAKE_OSX_DEPLOYMENT_TARGET is '10.10' but CMAKE_OSX_SYSROOT:
[cmake]
[cmake]    ""
[cmake]
[cmake]   is not set to a MacOSX SDK with a recognized version.  Either set
[cmake]   CMAKE_OSX_SYSROOT to a valid SDK or set CMAKE_OSX_DEPLOYMENT_TARGET to
[cmake]   empty.
[cmake] Call Stack (most recent call first):
[cmake]   Modules/CMakeSystemSpecificInformation.cmake:36 (include)
[cmake]   CMakeLists.txt:16 (project)

Unless the CMAKE_OSX_DEPLOYMENT_TARGET is set to empty. We set it to empty by
unsetting the MACOSX_DEPLOYMENT_TARGET environment variable.

Similar problem happens when cmake is used to build other packages. There the
fix is to unset the CMAKE_OSX_DEPLOYMENT_TARGET variable on the command line.