Update ncurses to 5.9-20141206
Upstream only distributes the 5.9 version plus the 20141206 patch in a form
that is not easy to apply, so I pushed it into a git repository.

The clang-fix.patch, even though it applies, actually makes clang fail to build
it and removing the patch fixes it. The package now builds on OS X 10.10 both
with the native clang as well as with Hashstack's gcc.